src/backend/requirements.txt
fastapi[standard]==0.116.1
sqlalchemy==2.0.43
pydantic-settings==2.10.1
pydantic==2.11.7
structlog==25.4.0
alembic==1.16.4
python-ulid[pydantic]==3.1.0
